feat(llm): FR-263 add Azure OpenAI provider
- Add azure provider via langchain-azure-ai AzureAIOpenAIApiChatModel
- Extract provider functions to llm_providers.py (file-size gate)
- Config: AZURE_AI_ENDPOINT, AZURE_AI_API_KEY, AZURE_MODEL env vars
- Tests: 9 unit tests covering creation, validation, caching, dispatch
- Update ARCHITECTURE.md provider count, CLAUDE.md env table
- Changelog fragment and diary reflection
