+++ .env.sample
@@ -34,8 +34,14 @@ GOOGLE_CLOUD_PROJECT=your-gcp-project-id
 GOOGLE_CLOUD_LOCATION=us-central1
 # VERTEX_MODEL=gemini-2.0-flash
 
+# Azure AI (optional - only needed if using provider: azure)
+# Requires: pip install yamlgraph[azure]
+AZURE_AI_ENDPOINT=https://my-resource.services.ai.azure.com/openai/v1
+AZURE_AI_API_KEY=your-azure-ai-api-key-here
+# AZURE_MODEL=gpt-4o
+
 # Default Provider Selection (optional)
-# Options: anthropic, google, inception, lmstudio, mistral, openai, replicate, vertex, xai
+# Options: anthropic, azure, deepseek, google, inception, lmstudio, mistral, openai, replicate, vertex, xai
 # Default: anthropic
 # PROVIDER=anthropic
