+++ changelog/unreleased/fr-263-azure-openai-provider.md
@@ -1,0 +1,6 @@
+---
+type: feat
+scope: llm
+req: REQ-YG-010
+---
+- **FR-263 Azure OpenAI Provider**: Add `azure` provider to LLM factory using `langchain-azure-ai`, enabling Azure-hosted models (GPT-4o, Llama, Mistral, Cohere) via Azure AI Foundry. (REQ-YG-010)
